+++ src/components/QuizOverview/styles.js
@@ -1,0 +1,8 @@
+
+import { StyleSheet } from 'react-native'
+
+import { common } from '../../styles'
+
+export const styles = StyleSheet.create({
+  centered : common.centeredContainer
+});