+++ src/views/QuizDetail.js
@@ -2,33 +2,31 @@ import React, { Component } from 'react'
 import { View, Text, TouchableOpacity } from 'react-native'
 import { connect } from 'react-redux'
 
-class QuizDetail extends Component {
+import QuizOverview from '../components/QuizOverview'
 
+class QuizOverviewDetail extends Component {
+
+	static navigationOptions = ({ navigation }) => {
+		const quizTitle = navigation.state.params.quiz.title
+
+		return {
+			title: quizTitle
+		}
+	}
+	
 	render(){
-
-		let {title, questions} = this.props.quiz
-		let cardNumber = questions.length
-		let numberText = !questions.length ? 'No cards' : questions.length > 1 ? 'Cards' : 'Card'
-
+		console.log('PROPS', this.props)		
+		// this should be refactored into a reducer later
+		let passedQuizOverview = this.props.navigation.state.params.quiz
 		return(
-			<View>
-				<Text>{title}</Text>
-				<Text>{cardNumber} {numberText}</Text>
-				<TouchableOpacity>
-					<Text>Add Card</Text>
-				</TouchableOpacity>
-				<TouchableOpacity>
-					<Text>Start Quiz</Text>
-				</TouchableOpacity>
-			</View>
+				<QuizOverview quiz={passedQuizOverview} />
 		)
 	}
 }
 
 function mapStateToProps({quizList}){
-
 	// will be modified to work with individual quizes instead of base constant
-	return {quiz: quizList[0]}
+	return {quizPlaceholder: quizList[0]}
 }
 
-export default connect(mapStateToProps)(QuizDetail)
+export default connect(mapStateToProps)(QuizOverviewDetail)